bug/minor + feat: tinymce code revamp
reorganize the tinymce code so templates also have the Image plugin and
handle dropped images
fix bug in admin panel where newcomer message was improperly displayed
