+++ src/templates/admin.html
@@ -186,7 +186,7 @@
         <!-- [html-validate-disable-block unique-landmark, element-permitted-content, no-deprecated-attr, prefer-native-element, no-unused-disable, element-permitted-content: suppress errors from tinymce] -->
       {%- endif %}
       <textarea style='height:400px' class='mceditable' name='newcomer_banner' id='newcomer_banner'>
-        {{ src.onboarding_email_admins_body|raw }}
+        {{ src.newcomer_banner|raw }}
       </textarea>
     </div>
     <div class='mt-2 text-center'>
