+++ src/ts/admin.ts
@@ -22,7 +22,6 @@ import { getEditor } from './Editor.class';
 import { Api } from './Apiv2.class';
 import { EntityType, Model, Action, Selected } from './interfaces';
 import tinymce from 'tinymce/tinymce';
-import { getTinymceBaseConfig } from './tinymce';
 import Tab from './Tab.class';
 
 document.addEventListener('DOMContentLoaded', () => {
@@ -36,11 +35,7 @@ document.addEventListener('DOMContentLoaded', () => {
   TabMenu.init(document.querySelector('.tabbed-menu'));
 
   const editor = getEditor();
-  if (editor.type === 'tiny') {
-    tinymce.init(getTinymceBaseConfig('admin'));
-  } else {
-    editor.init();
-  }
+  editor.init('admin');
 
   function collectSelectable(name: string) {
     const collected = [];
